devel/volk: update to 2.5.0
Summary:
This is intended to be used with comms/gnuradio 3.9, which I have WIP (almost ready for review/test) which I will most likely present during the WIP session at devsummit.

`cpu_features` is disabled because that upstream doesn't have code to support us (yet).

Test Plan: Passes testport, however the real test is with new comms/gnuradio. A wider call for testing will follow for that. This will only have a chance at landing when that port is ready.

Reviewers: tcberner, fluffy, arrowd, db, yuri

Subscribers: mat

Differential Revision: https://reviews.freebsd.org/D30700