python.mk: remove setuptools from RUN_DEPENDS
Summary:
Historically this was needed as a RUN_DEPENDS due to consumers using the included `pkg_resources`, which has been deprecated for some time now in favour of the `importlib` suite. (`pkg_resources` consumers are strongly advised to migrate to `importlib`.)

Additionally, due to environment pollution, this cannot be specified as a RUN_DEPENDS anywhere; consumers of consumers that need this functionality (setuptools_scm comes to mind) are to specify this in BUILD_DEPENDS only. Otherwise certain dependency trees are wrong, which can prevent tests from being available amongst other things.

Test Plan: exp-run

Reviewers: #python, #portmgr!

Subscribers: mat

Differential Revision: https://reviews.freebsd.org/D39303