net-p2p/c-lightning: hook dependencies for the clnrest plugin
The clnrest plugin is already installed but it depends on a bunch of
Python packages. Add an option to net-p2p/c-lightning to install and
depend on those packages.

Further, 3 of those packages were not in the FreeBSD ports. Add them:
* net-p2p/pyln-bolt7 Lightning Network BOLT7 implementation
* net-p2p/pyln-client Python client library for lightningd (net-p2p/c-lightning)
* net-p2p/pyln-proto Lightning Network protocol implementation

To enable the plugin, add
clnrest-port=
clnrest-protocol=
clnrest-host=
to /usr/local/etc/lightningd-bitcoin.conf, see
https://docs.corelightning.org/docs/rest

Differential Revision:	https://reviews.freebsd.org/D43863