editors/openoffice-4: Upgrade to version 4.1.15
Upgrade to version 4.1.15, which has bug and security fixes.

Release notes:
  https://cwiki.apache.org/confluence/display/OOOUSERS/AOO+4.1.15+Release+Notes

Since we use the ports version of expat instead of bundling it into
the openoffice-* packagea,
  CVE-2022-43680: "Use after free" fixed in libexpat
does not apply to our ports. This issue was fixed when we upgraded the
expat port.

Remove the point release from the suffix appended to the names of
the wrapper script and the program directory.  It is annoying when
the suffix changes and we'll never support the simultaneous
installation of multiple versions where these values differ.

MFH:		2024Q1 tree branch name you plan to merge to.>
Security:	CVE-2012-5639
Security:	CVE-2023-1183
Security:	CVE-2023-47804